Refactoring planning and control
