+++ pylot/planning/messages.py
@@ -2,21 +2,28 @@ import erdos
 
 
 class WaypointsMessage(erdos.Message):
-    """Message class to be used to send waypoints.
+    """Message class to be used to send waypoints. Optionally can also send
+    a target speed for each waypoint.
 
     Args:
         timestamp (:py:class:`erdos.timestamp.Timestamp`): The timestamp of
             the message.
         waypoints (list(:py:class:`~pylot.utils.Transform`), optional): List of
             waypoint transforms.
+        target_speeds (list(float)), optional): List of target speeds.
     """
-    def __init__(self, timestamp, waypoints):
+    def __init__(self, timestamp, waypoints, target_speeds=None):
         super(WaypointsMessage, self).__init__(timestamp, None)
         self.waypoints = waypoints
+        if target_speeds is not None:
+            assert len(target_speeds) == len(waypoints), \
+                "Length of target speeds must match length of waypoints"
+        self.target_speeds = target_speeds
 
     def __str__(self):
-        return 'WaypointMessage(timestamp: {}, waypoints: {}'.format(
-            self.timestamp, self.waypoints)
+        return \
+            'WaypointMessage(timestamp: {}, waypoints: {}, target speeds: {}'\
+            .format(self.timestamp, self.waypoints, self.target_speeds)
 
 
 class BehaviorMessage(erdos.Message):
